# src/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Striqo Football Platform API
  description: API Documentation for the Striqo Football Platform. This API provides programmatic access to users, teams, tournaments, matches, and more.
  version: 1.0.0
  contact:
    name: API Support
    email: dev@striqo.com
servers:
  - url: /
    description: Current Environment
  - url: /api/v1
    description: V1 APIs
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        error:
          type: string
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Invalid request parameters"
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
...
security:
  - bearerAuth: []
paths:
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password]
              properties:
                username: { type: string, minLength: 3 }
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Validation error or username/email taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...